+++ example/lib/main.dart
@@ -1,3 +1,4 @@
+import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:heart_overlay/heart_overlay.dart';
 
@@ -18,10 +19,23 @@ class MainApp extends StatelessWidget {
   }
 }
 
-class HeartOverlayTestWidget extends StatelessWidget {
+class HeartOverlayTestWidget extends StatefulWidget {
   const HeartOverlayTestWidget({
     super.key,
   });
+
+  @override
+  State<HeartOverlayTestWidget> createState() => _HeartOverlayTestWidgetState();
+}
+
+class _HeartOverlayTestWidgetState extends State<HeartOverlayTestWidget> {
+  late HeartOverlayController heartOverlayController;
+
+  @override
+  void initState() {
+    super.initState();
+    heartOverlayController = HeartOverlayController();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -36,11 +50,13 @@ class HeartOverlayTestWidget extends StatelessWidget {
         children: [
           Expanded(
             child: HeartOverlay(
+              controller: heartOverlayController,
               duration: const Duration(seconds: 5),
               backgroundColor: Colors.black,
               tapDownType: TapDownType.single,
               // verticalOffset: 20,
               // horizontalOffset: -100,
+              enableGestures: true,
               cacheExtent: 30,
               splashAnimationDetails: const SplashAnimationDetails(
                 enableSplash: true,
@@ -54,20 +70,51 @@ class HeartOverlayTestWidget extends StatelessWidget {
               onPressed: (numberOfHearts) {
                 // Do something with the number of hearts or do something whenever the icon appears
               },
-              // child: Center(
-              //   child: ElevatedButton(
-              //     onPressed: () {},
-              //     child: const Text('Tap here'),
-              //   ),
-              // ),
             ),
           ),
+          Padding(
+            padding: const EdgeInsets.all(50.0),
+            child: ElevatedButton(
+              onPressed: () {
+                // Get the center offset
+                // Offset centerOffset =
+                //     heartOverlayController.getCenterScreenOffset(context);
+
+                double iconSize = heartOverlayController.getSize() * 2;
+                double height = heartOverlayController.getHeight() ??
+                    MediaQuery.of(context).size.height;
+                double width = heartOverlayController.getWidth() ??
+                    MediaQuery.of(context).size.width;
+
+                Offset offset = Offset(
+                  // DX
+                  (Random().nextDouble() * width).clamp(
+                    iconSize,
+                    width - iconSize,
+                  ),
+                  // DY
+                  (Random().nextDouble() * height).clamp(
+                    iconSize,
+                    height - iconSize - 100,
+                  ), // - 100 is for the button padding
+                );
+
+                // offset += centerOffset; // You can add offsets
+
+                heartOverlayController.showIcon(
+                  offset: offset,
+                );
+              },
+              child: const Text('Show Icon'),
+            ),
+          )
           // Expanded(
           //   child: HeartOverlay(
           //     duration: const Duration(seconds: 2),
           //     backgroundColor: Colors.white10,
           //     tapDownType: TapDownType.single,
           //     icon: const Text('✌️'),
+          //     // Splash would not work because the [icon] provided is not of type [Icon]
           //     splashAnimationDetails: const SplashAnimationDetails(
           //       enableSplash: true,
           //     ),